Fix incorrect data types

+++ lab04_interval_scheduling/main.cpp
@@ -5,7 +5,7 @@
 int main() {
     int n;
     std::cin >> n;
-    std::vector<std::pair<long, long>> tasks(n);
+    std::vector<std::pair<long long, long long>> tasks(n);
     for (auto &task : tasks) {
         std::cin >> task.second;
     }
@@ -14,7 +14,7 @@ int main() {
     }
     sort(begin(tasks), end(tasks));
     auto end = -1;
-    auto count = 0;
+    auto count = 0ll;
     for (const auto &task : tasks) {
         if (task.second > end) {
             ++count;
